refactor(cli): extract secondary commands into cli_commands.go to reduce main.go size
